+++ app/src/main/java/com/twilio/video/app/data/Preferences.java
@@ -17,7 +17,9 @@
 package com.twilio.video.app.data;
 
 import com.twilio.video.AudioCodec;
+import com.twilio.video.OpusCodec;
 import com.twilio.video.VideoCodec;
+import com.twilio.video.Vp8Codec;
 import com.twilio.video.app.BuildConfig;
 
 public class Preferences {
@@ -40,9 +42,9 @@ public class Preferences {
     public static final String ENABLE_INSIGHTS = "pref_enable_insights";
     public static final boolean ENABLE_INSIGHTS_DEFAULT = true;
     public static final String VIDEO_CODEC = "pref_video_codecs";
-    public static final String VIDEO_CODEC_DEFAULT = VideoCodec.VP8.toString();
+    public static final String VIDEO_CODEC_DEFAULT = Vp8Codec.NAME;
     public static final String AUDIO_CODEC = "pref_audio_codecs";
-    public static final String AUDIO_CODEC_DEFAULT = AudioCodec.OPUS.toString();
+    public static final String AUDIO_CODEC_DEFAULT = OpusCodec.NAME;
     public static final String MAX_AUDIO_BITRATE = "pref_max_audio_bitrate";
     public static final int MAX_AUDIO_BITRATE_DEFAULT = 0;
     public static final String MAX_VIDEO_BITRATE = "pref_max_video_bitrate";